devtools/client/webide/test/test_fullscreenToolbox.html

<!DOCTYPE html>

<html>

  <head>
    <meta charset="utf8">
    <title></title>

    <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
    <script src="chrome://mochikit/content/chrome-harness.js"></script>
    <script type="application/javascript" src="head.js"></script>
    <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  </head>

  <body>

    <script type="application/javascript">
      function connectToLocal(win, docRuntime) {
        return new Promise(resolve => {
          win.AppManager.connection.once(
              win.Connection.Events.CONNECTED,
              resolve);
          docRuntime.querySelectorAll(".runtime-panel-item-other")[1].click();
        });
      }

      window.onload = function() {
        SimpleTest.waitForExplicitFinish();

        (async function() {
          const win = await openWebIDE();
          const docProject = getProjectDocument(win);
          const docRuntime = getRuntimeDocument(win);
          win.AppManager.update("runtime-list");

          const onGlobalActors = waitForUpdate(win, "runtime-global-actors");
          const onRuntimeTargets = waitForUpdate(win, "runtime-targets");
          connectToLocal(win, docRuntime);
          await onGlobalActors;
          await onRuntimeTargets;

          // Select main process
          SimpleTest.executeSoon(() => {
            docProject.querySelectorAll("#project-panel-runtimeapps .panel-item")[0].click();
          });

          await waitForUpdate(win, "project");

          ok(win.UI.toolboxPromise, "Toolbox promise exists");
          await win.UI.toolboxPromise;

          const nbox = win.document.getElementById("containerbox");
          ok(!nbox.hasAttribute("toolboxfullscreen"), "Toolbox is not fullscreen");

          win.Cmds.showRuntimeDetails();

          ok(!nbox.hasAttribute("toolboxfullscreen"), "Toolbox is not fullscreen");

          await win.Cmds.disconnectRuntime();

          await closeWebIDE(win);

          DebuggerServer.destroy();

          SimpleTest.finish();
        })();
      };
    </script>
  </body>
</html>
